Created a Report screen to display the 4 report queries.
Finished tha AllCars to display all the contents of the car table with a refresh button.

Created a Sold class to display all the contents of the sold table with a refresh button.

Created a Service class modeled after the inventory class to display serviced cars. Created the AddAServiceCar class to allow a car to be added to the service table.

Added a few prepared statements to the DBConnection class.